fix readDouble nextLine error

+++ src/s1_02_exceptions_lvl1/Product.java
@@ -2,22 +2,22 @@ package s1_02_exceptions_lvl1;
 
 public class Product {
     private String name;
-    private float price;
+    private double price;
 
-    public Product(String name, float price) {
+    public Product(String name, double price) {
         this.name = name;
         this.price = price;
     }
     public String getName() {
         return this.name;
     }
-    public float getPrice() {
+    public double getPrice() {
         return this.price;
     }
     public void setName(String name) {
         this.name = name;
     }
-    public void setPrice(float price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 }
